pyproject.toml
[project]
name = "ai-app"
version = "0.1.0"
description = ""
authors = [
    {name = "ofelipechan",email = "felipe_s.c1@hotmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "langchain-core (>=0.3.68,<0.4.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "unstructured[md] (>=0.18.5,<0.19.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "chromadb (>=1.0.15,<2.0.0)",
    "langchain-chroma (>=0.2.4,<0.3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "3.13.5"